refactor 04
