+++ packages/apostrophe/modules/@apostrophecms/ai/index.js
@@ -7,7 +7,6 @@
 // name the default with `options.provider`.
 
 const Ajv = require('ajv/dist/2020').default;
-const { FINISH_REASONS } = require('./lib/constants');
 const {
   isObject, isAbort, startupFail
 } = require('./lib/util');
@@ -89,6 +88,8 @@ module.exports = {
     return {
       ...require('./lib/startup')(self),
       ...require('./lib/normalize')(self),
+      ...require('./lib/request')(self),
+      ...require('./lib/adapter-call')(self),
       ...require('./lib/aspect')(self),
 
       // Register a provider adapter. Adapters self-register in their own
@@ -102,61 +103,6 @@ module.exports = {
       },
       getAdapter(name) {
         return self.adapters[name];
-      },
-      // Resolve a call's routing options to a concrete routing entry,
-      // with the same precedence generate will use: explicit
-      // provider+model, else the call's effort level, else the default
-      // level. Throws "invalid" on unresolvable calls; unknown models
-      // are not an error.
-      //
-      // Options:
-      // `provider`, `model` (strings, only together): the explicit
-      //   target, bypassing the routing table;
-      // `effort` (string): the routing level to resolve;
-      // `capability` (only 'image'): resolve the image route instead of
-      //   the effort table;
-      // `reasoning` (string): override the resolved entry's reasoning.
-      resolve(options = {}) {
-        const invalid = (message) => {
-          throw self.apos.error('invalid', message);
-        };
-        const {
-          provider, model, effort, capability, reasoning
-        } = options;
-
-        if (capability !== undefined && capability !== 'image') {
-          invalid(`unknown capability "${capability}"`);
-        }
-        if (provider || model) {
-          if (!provider || !model) {
-            invalid('"provider" and "model" must be given together');
-          }
-          if (!self.providers[provider]) {
-            invalid(`"${provider}" is not a configured provider`);
-          }
-          return {
-            provider,
-            model,
-            ...(reasoning !== undefined && { reasoning })
-          };
-        }
-        let row;
-        if (capability === 'image') {
-          if (!self.options.image) {
-            invalid('no "image" route is configured');
-          }
-          row = self.options.image;
-        } else {
-          const level = effort || self.effortDefault;
-          row = self.effortTable[level];
-          if (!row) {
-            invalid(`effort level "${level}" resolves to no routing entry`);
-          }
-        }
-        return {
-          ...row,
-          ...(reasoning !== undefined && { reasoning })
-        };
       },
       // Synchronous introspection: the model a call with these options
       // would hit and what it offers. Accepts the same options as
@@ -350,76 +296,6 @@ module.exports = {
             .map((error) => `${error.path}: ${error.message || error.name}`)
             .join('; ');
         }
-      },
-      // Assemble the normalized adapter request from `options`, a
-      // canonical object as produced by normalizeGenerateOptions:
-      // resolve routing, default maxTokens to the model's declared
-      // output ceiling when it is known, translate the cache level to
-      // the { ttl } policy. Returns { provider, request }: the resolved
-      // provider name and the request handed to its adapter —
-      // { system?, messages, tools?, schema?, model, maxTokens?,
-      // reasoning?, cache: false | { ttl }, signal? }, optional fields
-      // present only when they resolved to a value. Request tools carry
-      // only { name, description, input } — handlers and result schemas
-      // never reach an adapter; a structured-output `schema` (JSON
-      // Schema) passes through for the adapter to place on its
-      // provider's native structured mode.
-      buildRequest(options) {
-        const info = self.modelInfo({
-          provider: options.provider,
-          model: options.model,
-          effort: options.effort,
-          reasoning: options.reasoning
-        });
-        const maxTokens = options.maxTokens ?? info.maxOutputTokens;
-        return {
-          provider: info.provider,
-          request: {
-            ...(options.system !== undefined && { system: options.system }),
-            messages: options.messages,
-            ...(options.tools.length && { tools: self.wireTools(options.tools) }),
-            ...(options.schema !== undefined && { schema: options.schema }),
-            model: info.model,
-            ...(maxTokens !== undefined && { maxTokens }),
-            ...(info.reasoning !== undefined && { reasoning: info.reasoning }),
-            cache: options.cache === false
-              ? false
-              : { ttl: options.cache },
-            ...(options.signal !== undefined && { signal: options.signal })
-          }
-        };
-      },
-      // Assemble the adapter request without routing, for mock mode
-      // with no providers configured. Same input and return shape as
-      // buildRequest, with the call's explicit provider and model when
-      // given and "mock" placeholders otherwise; no model metadata
-      // exists here, so maxTokens appears only when the call sets it.
-      buildMockRequest(options) {
-        return {
-          provider: options.provider ?? 'mock',
-          request: {
-            ...(options.system !== undefined && { system: options.system }),
-            messages: options.messages,
-            ...(options.tools.length && { tools: self.wireTools(options.tools) }),
-            ...(options.schema !== undefined && { schema: options.schema }),
-            model: options.model ?? 'mock',
-            ...(options.maxTokens !== undefined && { maxTokens: options.maxTokens }),
-            ...(options.reasoning !== undefined && { reasoning: options.reasoning }),
-            cache: options.cache === false
-              ? false
-              : { ttl: options.cache },
-            ...(options.signal !== undefined && { signal: options.signal })
-          }
-        };
-      },
-      // The model-facing face of activated tool definitions, as placed
-      // on the adapter request
-      wireTools(tools) {
-        return tools.map((tool) => ({
-          name: tool.name,
-          description: tool.description,
-          input: tool.input
-        }));
       },
       // The built-in mock standing in for every adapter chat under
       // APOS_AI_MOCK. Consults the "mock" option first when the module
@@ -1282,266 +1158,6 @@ module.exports = {
           await run(call, index);
         }
         return outcomes;
-      },
-      // Throw "invalid" when the configured provider named by
-      // `provider` does not declare `capability` (a key of the
-      // capabilities map, e.g. 'text'). A call needing a capability the
-      // routed provider lacks is a clear error, never a silent
-      // re-route.
-      checkCapability(provider, capability) {
-        if (!self.providers[provider]?.capabilities?.[capability]) {
-          throw self.apos.error(
-            'invalid',
-            `provider "${provider}" does not declare the "${capability}" capability`
-          );
-        }
-      },
-      // Run one adapter call with retries. `req` is the caller's
-      // request, enriching the failure records; `record` is an
-      // activated entry of `self.providers` (supplies the adapter and
-      // its normalizeError); `request` is the normalized adapter
-      // request, read only for record context; `call` is an async thunk
-      // performing a single adapter call and validating its response.
-      // Resolves with the thunk's value; throws normalized apos errors
-      // — every throw is routed through the adapter's required
-      // normalizeError, the core reacts on codes only, and only the
-      // transient code is retried, waiting per retryDelay under the
-      // retryAttempts and retryMaxElapsed budgets. `call` is retried
-      // whole, so response validation belongs inside it: a truncated
-      // body must travel the same retry path.
-      //
-      // A normalized error may carry hints in `error.data`, the only
-      // properties the engine reads — all optional, attached by the
-      // adapter's normalizeError:
-      // `status` (integer): the provider's HTTP status code;
-      // `kind` (string, on the transient code): which transient
-      //   failure this is — 'rateLimit', 'overload', 'timeout' or
-      //   'network';
-      // `retryAfter` (number, in SECONDS): the provider's Retry-After;
-      //   replaces the computed backoff delay (see retryDelay);
-      // `requestId` (string): the provider's request id, for support.
-      // Hints shape the delay and the records, never the routing: the
-      // error's code alone decides retry versus stop. All of these
-      // are written to the failure and retry log records — treat
-      // `error.data` as log-bound and never put sensitive data (keys,
-      // credentials, personal data) in it.
-      //
-      // Every failure and every retry decision emits one structured log
-      // record: type `retry` (warn) when the call will be retried, type
-      // `failure` (error) when it stops, with { provider, model, code,
-      // status, kind, requestId, retryAfter, attempt, elapsed, action,
-      // reason?, delay? } — enough to tell a rate limit from an
-      // overload, a timeout or bad config from the one record. A
-      // `failure` record also carries the stack of the original throw.
-      async callAdapter(req, record, request, call) {
-        const started = Date.now();
-        for (let attempt = 1; ; attempt++) {
-          try {
-            return await call();
-          } catch (e) {
-            // A fired abort signal is the caller's own doing, not a
-            // provider failure: no failure record, no retry
-            if (isAbort(e)) {
-              throw e;
-            }
-            const error = (e?.aposError ? e : record.adapter.normalizeError(e)) || e;
-            const elapsed = Date.now() - started;
-            const data = {
-              provider: record.name,
-              model: request.model,
-              code: error.name,
-              status: error.data?.status,
-              kind: error.data?.kind,
-              requestId: error.data?.requestId,
-              retryAfter: error.data?.retryAfter,
-              attempt,
-              elapsed
-            };
-            // The original throw site is the useful trace when the
-            // adapter wrapped a client error
-            const stack = e?.stack || error.stack;
-            if (error.name !== 'aiRetry') {
-              self.logError(req, 'failure', error.message, {
-                ...data,
-                action: 'stop',
-                stack
-              });
-              throw error;
-            }
-            if (attempt >= self.options.retryAttempts) {
-              self.logError(req, 'failure', error.message, {
-                ...data,
-                action: 'stop',
-                reason: 'attempts',
-                stack
-              });
-              throw error;
-            }
-            const delay = self.retryDelay(attempt, error);
-            if (elapsed + delay > self.options.retryMaxElapsed) {
-              self.logError(req, 'failure', error.message, {
-                ...data,
-                action: 'stop',
-                reason: 'budget',
-                delay,
-                stack
-              });
-              throw error;
-            }
-            self.logWarn(req, 'retry', error.message, {
-              ...data,
-              action: 'retry',
-              delay
-            });
-            await self.pause(delay);
-          }
-        }
-      },
-      // The wait in milliseconds before the attempt following `attempt`
-      // (1-based), after the normalized transient failure `error`: the
-      // provider's Retry-After (seconds, in the error's
-      // `data.retryAfter`) when it sent one, else exponential backoff
-      // with jitter — retryBaseDelay * 2^(attempt - 1), scaled by a
-      // random factor in [1, 2) so synchronized clients spread out.
-      retryDelay(attempt, error) {
-        const retryAfter = error.data?.retryAfter;
-        if (Number.isFinite(retryAfter) && retryAfter >= 0) {
-          return retryAfter * 1000;
-        }
-        const curve = self.options.retryBaseDelay * Math.pow(2, attempt - 1);
-        return Math.floor(curve * (1 + Math.random()));
-      },
-      // Wait `ms` before the next attempt; a separate method so tests
-      // can observe or skip real waiting
-      pause(ms) {
-        return new Promise(resolve => setTimeout(resolve, ms));
-      },
-      // Enforce the assistant-turn contract on `turn`, an adapter chat
-      // response { content, finishReason, usage, model? }, and return
-      // it unchanged. A missing or unknown finishReason, or malformed
-      // content or usage, is a truncated or malformed response: it
-      // throws the transient code so the call travels the retry path —
-      // never a shorter-than-intended "success".
-      validateTurn(turn) {
-        const malformed = (detail) => {
-          throw self.apos.error('aiRetry', `malformed assistant turn: ${detail}`);
-        };
-        if (!isObject(turn)) {
-          malformed('not an object');
-        }
-        if (!Array.isArray(turn.content)) {
-          malformed('"content" must be an array of content parts');
-        }
-        for (const part of turn.content) {
-          if (!isObject(part) || typeof part.type !== 'string') {
-            malformed('content parts must be objects with a "type"');
-          }
-          if (part.type === 'text' && typeof part.text !== 'string') {
-            malformed('text parts must carry a string "text"');
-          }
-          if (part.type === 'toolCall' && (
-            typeof part.id !== 'string' || !part.id ||
-            typeof part.name !== 'string' || !isObject(part.input)
-          )) {
-            malformed('toolCall parts must carry a string "id" and "name" and an object "input"');
-          }
-        }
-        if (!FINISH_REASONS.includes(turn.finishReason)) {
-          malformed(`"${turn.finishReason}" is not a finish reason`);
-        }
-        if (turn.finishReason === 'toolCalls' &&
-          !turn.content.some(part => part.type === 'toolCall')) {
-          malformed('a "toolCalls" finish reason without toolCall parts');
-        }
-        if (!isObject(turn.usage) ||
-          !Number.isFinite(turn.usage.inputTokens) ||
-          !Number.isFinite(turn.usage.outputTokens)) {
-          malformed('"usage" must carry inputTokens and outputTokens');
-        }
-        return turn;
-      },
-      // Enforce the adapter image contract on `result` ({ images,
-      // model?, usage?, size? }) and return it unchanged. A missing or
-      // empty image list, or a malformed image, is a malformed
-      // response: it throws the transient code so the call travels
-      // the retry path — a model whim never returns an empty success.
-      validateImageResult(result) {
-        const malformed = (detail) => {
-          throw self.apos.error('aiRetry', `malformed image result: ${detail}`);
-        };
-        if (!isObject(result)) {
-          malformed('not an object');
-        }
-        if (!Array.isArray(result.images) || !result.images.length) {
-          malformed('"images" must be a non-empty array');
-        }
-        for (const image of result.images) {
-          if (!isObject(image) || typeof image.type !== 'string' ||
-            typeof image.data !== 'string' || !image.data) {
-            malformed('images must carry a string "type" and a non-empty "data"');
-          }
-        }
-        if (result.usage !== undefined && !isObject(result.usage)) {
-          malformed('"usage" must be an object');
-        }
-        return result;
-      },
-      // The anti-hallucination backstop for structured output. The
-      // adapter has already extracted the final answer into `turn.object`
-      // from its provider's structured mode (an adapter concern,
-      // since only the dialect knows where the object lives); this
-      // validates it against `validate`, the compiled validator for the
-      // call's `schema` (normalizeGenerateOptions). The provider's
-      // native mode does the real work — this only catches a stray
-      // non-conforming or missing response. A missing object or a schema
-      // mismatch is a malformed model response: like a malformed turn it
-      // throws the transient code so the call travels the retry path.
-      validateStructured(turn, validate) {
-        if (turn.object === undefined) {
-          throw self.apos.error('aiRetry', 'the provider returned no structured output');
-        }
-        if (!validate(turn.object)) {
-          throw self.apos.error('aiRetry', `structured output does not match the schema: ${self.ajv.errorsText(validate.errors, { dataVar: 'object' })}`);
-        }
-      },
-      // Build generate's unified return object from `context` (the
-      // event payload carrying the provider name and the live request,
-      // whose messages already include every appended turn), `turn`
-      // (the final validated adapter response) and the loop's
-      // accumulations: executed `steps`, `usage` aggregated across
-      // every model turn, the `pending` tool calls when the step
-      // budget cut the loop, the validated structured `object` when the
-      // call passed a `schema`, and whether the call had tools at all
-      // (`steps` appears only then). Returns { text, messages,
-      // finishReason, usage, model, provider } plus `object`, `steps`
-      // and `toolCalls` as described on generate; which fields are
-      // populated tells the caller what happened. The transcript is
-      // resumable as the next call's `messages`.
-      assembleResult(context, turn, {
-        steps, usage, pending, object, hadTools, finishReason
-      }) {
-        // `turn` is null when a cancellation aborted the first provider
-        // call of a step: there is no final assistant turn, only the
-        // work recorded so far
-        const text = (turn ? turn.content : [])
-          .filter(part => part.type === 'text')
-          .map(part => part.text)
-          .join('');
-        return {
-          text,
-          ...(object !== undefined && { object }),
-          messages: [ ...context.request.messages ],
-          ...(hadTools && { steps }),
-          ...(pending && { toolCalls: pending }),
-          // The step budget cutting the loop is its own finish reason,
-          // like the token budget's 'length'; an explicit override wins
-          // (a cancelled run)
-          finishReason: finishReason ||
-            (pending ? 'maxSteps' : turn.finishReason),
-          usage,
-          model: turn?.model || context.request.model,
-          provider: context.provider
-        };
       }
     };
   }
